requirements.txt
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
idna==3.4
instagrapi==1.17.0
Pillow==9.4.0
praw==7.7.0
prawcore==2.3.0
pycodestyle==2.10.0
pycryptodomex==3.17
pydantic==1.10.5
PySocks==1.7.1
python-dotenv==1.0.0
requests==2.28.2
typing_extensions==4.5.0
update-checker==0.18.0
urllib3==1.26.15
websocket-client==1.5.1
